Refactoring command execution functionality
This commits overwrites some of the functionality related to executing
non built-in commands. It substitutes function pointers for finding
compiled binaries on the associated bin/ folder specified on the PATH
env variable. It also modifies Makefile compilation as it does not link
"ls", "grep" nor "chmod" command files to the main file "tesh". It keeps
commands separated so that main file is not recompile every time a new
command is added.